!165 fix: use newer buddy alloc and bump tustc to '1.95.0'
* fix: replace magic number with constant for page size in KernelSlabPool
* fix: add missing target for cargo dependency prefetching
* fix: simplify match statements in NetDeviceOps and Pollable implementations
* fix: simplify match statements in EthernetDevice and VsockStreamTransport
* fix: remove unused cfg_select feature
* fix: update rust-analyzer features and remove unused feature flags
* fix: update rust-analyzer features in settings.json
* fix: update buddy-slab-allocator to use workspace instead of version
* fix: use newer buddy alloc and bump tustc to '1.95.0'

+++ fs/rsext4/src/ext4.rs
@@ -1370,7 +1370,7 @@ fn build_uninit_group_desc(
     desc.bg_inode_table_lo = gl.group_inode_table_startblocks as u32;
 
     // 理论空闲块数：整组减去元数据块
-    let used_meta = gl.metadata_blocks_in_group as u32;
+    let used_meta = gl.metadata_blocks_in_group;
     let free_blocks = layout.blocks_per_group.saturating_sub(used_meta);
 
     if group_id == 0 {
